Parent-Child Relationship Management
(Transform is not implemented yet)
-------------------------------------
Encountered a big question:
vector or unordered set for the children container? - CHOSE VECTOR.
Added many functions for children management for adding, removing, accessing and most importantly - SetParent().

+++ GameEngine/Scene.cpp
@@ -103,7 +103,7 @@ void Scene::RemoveObjectByName(const std::string& goName)
 	}
 }
 
-void Scene::RemoveObjectByID(const size_t index)
+void Scene::RemoveObjectByID(const unsigned int index)
 {
 	if (index >= m_Objects.size())
 		return;
